Refactor: Rename app - expense to fyle_expense - task to task_log
