+++ apps/fyle_expense/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ExpenseConfig(AppConfig):
-    name = 'expense'
+    name = 'fyle_expense'
