+++ apps/task_log/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class TaskConfig(AppConfig):
-    name = 'task'
+    name = 'task_log'
