+++ apps/task_log/urls.py
@@ -3,7 +3,7 @@ task app URL Configuration
 """
 from django.urls import path
 
-from apps.task.views import TaskLogView, TaskLogDetailsView
+from apps.task_log.views import TaskLogView, TaskLogDetailsView
 
 urlpatterns = [
     path('', TaskLogView.as_view(), name="tasks"),
